requirements.txt
pymupdf
numpy
Pillow
pyserial
opencv-python
digi-xbee

# emisor es COM4y receptor es COM3